+++ CHANGELOG.md
@@ -13,12 +13,14 @@ This changelog is effective from the 2025 releases.
 
 ### Added
 * MultiJob now supports generic Job types for the self.children attribute
+* Function `view_orbital` to visualize orbitals of completed AMS jobs 
 
 ### Changed
 * `Molecule.readmol2` can now read non-integer bond orders
 * `Molecule.writepdb` now writes default values for res, resnum, fix, and occ columns
 * `view` function uses stdin mode for AMSview, reducing overhead for image creation
 * `vasp_output_to_ams` now reads the MD time step from `POTIM` in the OUTCAR and labels molecular-dynamics runs (`IBRION = 0`) as such, instead of using a fixed default time step
+* `view` function supports AMSJob and rkf as inputs
 
 ### Fixed
 * Xvfb backend for `view` function with 0 display value